pyproject.toml
[project]
name = "iotdb_mcp_server"
packages = [{ include = "src/iotdb_mcp_server" }]
version = "0.1.0"
description = "A Model Context Protocol (MCP) server that enables secure interaction with GreptimeDB databases. This server allows AI assistants to list tables, read data, and execute SQL queries through a controlled interface, making database exploration and analysis safer and more structured."
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "apache-iotdb>=2.0.1b0",
    "pyyaml>=6.0.2",
]
[[project.authors]]
name = "Yuan Tian"
email = "jackietien97@gmail.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/iotdb_mcp_server/**/*.py",
]

[tool.hatch.build.targets.wheel]
packages = [
    "src"
]

[project.scripts]
iotdb_mcp_server = "iotdb:main"